% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_energy.R
\name{calculate_stasiun_kpi}
\alias{calculate_stasiun_kpi}
\title{Hitung Kinerja Operasional Per Stasiun}
\usage{
calculate_stasiun_kpi(data)
}
\arguments{
\item{data}{Data.frame log transaksi hasil dari generate_ev_data().}
}
\value{
Sebuah data.frame ringkasan statistik per stasiun.
}
\description{
Mengklasifikasikan data transaksi untuk memantau total energi, sesi transaksi, dan finansial.
}
