% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_operator.R
\name{generate_ev_data}
\alias{generate_ev_data}
\title{Generate Data Simulasi Transaksi EV}
\usage{
generate_ev_data(bulan_tahun = "2026-06", seed = 2026)
}
\arguments{
\item{bulan_tahun}{String format "YYYY-MM" (default = "2026-06").}

\item{seed}{Angka acuan untuk generator bilangan acak (default = 2026).}
}
\value{
Sebuah data.frame berisi log transaksi pengisian EV.
}
\description{
Fungsi untuk menghasilkan data tiruan transaksi pengisian daya listrik sesuai dengan skenario stasiun RANA.
}
