% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_operator.R
\name{plot_energy_trend}
\alias{plot_energy_trend}
\title{Plot Tren Konsumsi Energi Harian}
\usage{
plot_energy_trend(data)
}
\arguments{
\item{data}{Data.frame log transaksi EV.}
}
\value{
Objek grafik ggplot2.
}
\description{
Menghasilkan grafik tren runtun waktu untuk kebutuhan Dashboard Monitoring Operator.
}
